test fix

+++ qendpoint-core/src/main/java/com/the_qa_company/qendpoint/core/util/disk/LongArray.java
@@ -238,7 +238,7 @@ public interface LongArray extends Iterable<Long> {
 		return EMPTY_ARRAY;
 	}
 
-	default int getEstimatedLocationArrayBucketSize() {
+	default long getEstimatedLocationArrayBucketSize() {
 		return 65536;
 	}
