+++ qendpoint-core/src/main/java/com/the_qa_company/qendpoint/core/util/disk/SyncLongArray.java
@@ -85,7 +85,7 @@ public class SyncLongArray implements LongArray {
 	}
 
 	@Override
-	public int getEstimatedLocationArrayBucketSize() {
+	public long getEstimatedLocationArrayBucketSize() {
 		return array.getEstimatedLocationArrayBucketSize();
 	}
